[DEL/#75] 더미 삭제
