+++ app/src/main/java/com/napzak/market/presentation/home/component/HomePopularItemGroup.kt
@@ -127,6 +127,7 @@ private fun HomePopularGroupPreview() {
                             isInterested = false,
                             tradeType = "팔아요",
                             tradeStatus = "판매중",
+                            isOwnedByCurrentUser = false,
                             isPriceNegotiable = false,
                         )
                     )
