+++ app/src/main/java/com/napzak/market/presentation/home/component/HomeRecommendationItemGroup.kt
@@ -114,6 +114,7 @@ private fun HomeRecommendationGroupPreview() {
                             isInterested = false,
                             tradeType = "구해요",
                             tradeStatus = "판매중",
+                            isOwnedByCurrentUser = false,
                             isPriceNegotiable = false,
                         )
                     )
