+++ app/src/main/java/com/napzak/market/presentation/home/state/HomeUiState.kt
@@ -1,7 +1,6 @@
 package com.napzak.market.presentation.home.state
 
 import com.napzak.market.core.common.state.UiState
-import com.napzak.market.core.type.TradeType
 import com.napzak.market.domain.home.model.HomeBanner
 import com.napzak.market.domain.home.model.ProductItem
 
@@ -11,59 +10,4 @@ data class HomeUiState(
     val recommendedItems: UiState<List<ProductItem>> = UiState.Loading,
     val popularItems: UiState<List<ProductItem>> = UiState.Loading,
     val searchItems: UiState<List<ProductItem>> = UiState.Loading,
-) {
-    companion object {
-        val dummyData = UiState.Success(
-            listOf(
-                ProductItem(
-                    productId = 1,
-                    productName = "딸기 마이멜로디 마스코트 인형",
-                    genreName = "산리오",
-                    price = 35000,
-                    uploadTime = "1시간전",
-                    photo = "",
-                    isInterested = false,
-                    tradeType = TradeType.SELL.name,
-                    tradeStatus = "판매중",
-                    isPriceNegotiable = false,
-                ),
-                ProductItem(
-                    productId = 2,
-                    productName = "딸기 마이멜로디 마스코트 인형",
-                    genreName = "산리오",
-                    price = 35000,
-                    uploadTime = "1시간전",
-                    photo = "",
-                    isInterested = false,
-                    tradeType = TradeType.SELL.name,
-                    tradeStatus = "판매중",
-                    isPriceNegotiable = false,
-                ),
-                ProductItem(
-                    productId = 3,
-                    productName = "딸기 마이멜로디 마스코트 인형",
-                    genreName = "산리오",
-                    price = 35000,
-                    uploadTime = "1시간전",
-                    photo = "",
-                    isInterested = false,
-                    tradeType = TradeType.SELL.name,
-                    tradeStatus = "판매중",
-                    isPriceNegotiable = false,
-                ),
-                ProductItem(
-                    productId = 4,
-                    productName = "딸기 마이멜로디 마스코트 인형",
-                    genreName = "산리오",
-                    price = 35000,
-                    uploadTime = "1시간전",
-                    photo = "",
-                    isInterested = false,
-                    tradeType = TradeType.SELL.name,
-                    tradeStatus = "판매중",
-                    isPriceNegotiable = false,
-                ),
-            )
-        )
-    }
-}
+)
